README.es: add a link to README.en
This is a one-way link because README.es.md is symlinked to by README.md;
this way a user reading the main project page on GitHub will see the docs
are in both languages.  A user who has managed to navigate to README.en.md
already realizes that there's also Spanish docs.

(but also I don't know how to write that message in Spanish)