Makefile, doc: update rule for modern Pandoc
Somewhere along the way, Pandoc turned on the "smart" markdown syntax
extension by default.  This breaks many of the man pages; "--flag" becomes
"–flag" (en-dash).  So add a flag to turn off that extension.

While we're at it, re-generate all of the docs, to verify that everything
still appears the way we expect it to with new pandoc/new flags.