lvpn-install: Work correctly with recent versions of systemd
I'm not sure when it switched, but recent versions of systemd won't accept
a plain file in /etc/systemd/system/*.wants/; it *must* be a symlink.  The
correct thing to do here is place the real file in /etc/systemd/system/,
and then create a symlink to it in /etc/systemd/system/multi-user.wants/.

This is why parabola_nu (proton.parabola.nu) hadn't been on LibreVPN for a
while; a systemd update made it start ignoring the installed
tincd@lvpn.service.