hosts/mapi: Fix syntax error
The `mapi` host file contains am "empty" line that actually has 2 spaces.
I don't know if previous versions of tinc were more lax, but as of
tinc-1.0.33, this is an error:

    No value for variable `' on line 8 while reading config file /etc/tinc/lvpn/hosts/mapi

The "=" is optional (if there's whitespace), and it trims leading+trailing
whitespace on the value, so it interprets the line as setting the ""
variable to an empty value.